% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.appears.plot.R
\name{char.appears.plot}
\alias{char.appears.plot}
\title{Function plots the presence of MLP:FiM characters in episodes of each season of the series}
\usage{
char.appears.plot(char.appears.df, col.palette)
}
\arguments{
\item{char.appears.df}{VA data frame containing information on the percentage of presence of characters in MLP:FiM seasons}

\item{col.palette}{color palette in the chart, a vector containing strings with color values stored in RGB hexadecimal, having the same length as the number of characters}
}
\value{
graph of percentage attendance of characters in MLP:FiM series
}
\description{
Function plots the presence of MLP:FiM characters in episodes of each season of the series
}
\examples{
char.appears.plot(char.appears.df, col.palette)
}
