% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.appears.R
\name{char.appears}
\alias{char.appears}
\title{Function calculates the percentage of characters in each season of the MLP:FiM series}
\usage{
char.appears(char.names, webpage)
}
\arguments{
\item{char.names}{Vector containing characters names}

\item{webpage}{String with page source code}
}
\value{
A data frame, the first column of which contains season numbers, while the subsequent ones contain the percentage of episodes in which particular characters appeared within a given season
}
\description{
Function calculates the percentage of characters in each season of the MLP:FiM series
}
\examples{
char.names <- c("Twilight Sparkle", "Rainbow Dash")
webpage <- webpage()
char.appears(char.names, webpage)
}
