% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.appears.single.R
\name{char.appears.single}
\alias{char.appears.single}
\title{Counting the percentage of the hero in each season}
\usage{
char.appears.single(char.name, webpage, lines.seasons, lines.endtab)
}
\arguments{
\item{char.name}{A single string containing the character's name}

\item{webpage}{String with page source code}

\item{lines.seasons}{A vector with the line numbers in the site's code that begin the sections about each season}

\item{lines.endtab}{A vector with line numbers in the page code containing the end of the arrays}
}
\value{
A vector containing the percentage of episodes in which a character appeared in each season
}
\description{
Counting the percentage of the hero in each season
}
\examples{
char.name <- c("Twilight Sparkle")
webpage <- webpage()
char.appears.single(char.name, webpage, lines.seasons, lines.endtab)
}
