% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.colors.R
\name{char.colors}
\alias{char.colors}
\title{Function returns a vector containing information about the coat color of MLP:FiM characters}
\usage{
char.colors(char.names)
}
\arguments{
\item{char.names}{A vector with characters names}
}
\value{
Function returns a graph containing the RGB values of the coat colors of the MLP:FiM characters
}
\description{
Function returns a vector containing information about the coat color of MLP:FiM characters
}
\examples{
char.colors(c("Twilight Sparkle", "Rarity"))
}
