% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.webpage.R
\name{get.webpage}
\alias{get.webpage}
\title{Downloading and reading a page from the fan wiki about
the appearances of the main characters in the series
"My Little Pony: Friendship is Magic"}
\usage{
get.webpage()
}
\value{
A html code.
}
\description{
Downloading and reading a page from the fan wiki about
the appearances of the main characters in the series
"My Little Pony: Friendship is Magic"
}
\examples{
webpage()
}
